# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZapCentral - Multi-tenant WhatsApp customer service platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, suspended, all]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantList"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid input
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by ID
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Tenant not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant settings
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Deactivate a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tenant deactivated
  /tenants/{tenantId}/stats:
    get:
      operationId: getTenantStats
      tags: [tenants]
      summary: Get tenant statistics summary
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantStats"

  # --- Users within tenant ---
  /tenants/{tenantId}/users:
    get:
      operationId: listTenantUsers
      tags: [users]
      summary: List users in a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantUser"
  /tenants/{tenantId}/users/invite:
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a user to the tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInviteInput"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "400":
          description: Invalid input
  /tenants/{tenantId}/users/{userId}:
    patch:
      operationId: updateTenantUser
      tags: [users]
      summary: Update user role or status in tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
    delete:
      operationId: removeTenantUser
      tags: [users]
      summary: Remove a user from the tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User removed

  # --- Departments ---
...
components:
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        planType:
          type: string
          enum: [trial, starter, professional, enterprise]
        status:
          type: string
          enum: [active, suspended, pending]
        maxAgents:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - planType
        - status
        - maxAgents
        - createdAt
        - updatedAt
    TenantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        slug:
          type: string
          minLength: 2
        logoUrl:
          type: string
        planType:
          type: string
          enum: [trial, starter, professional, enterprise]
        maxAgents:
          type: number
      required:
        - name
        - slug
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        logoUrl:
          type: ["string", "null"]
        planType:
          type: string
          enum: [trial, starter, professional, enterprise]
        status:
          type: string
          enum: [active, suspended, pending]
        maxAgents:
          type: number
    TenantList:
      type: object
      properties:
        tenants:
          type: array
          items:
            $ref: "#/components/schemas/Tenant"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - tenants
        - total
        - page
        - limit
    TenantStats:
      type: object
      properties:
        totalUsers:
          type: number
        activeAgents:
          type: number
        totalDepartments:
          type: number
        activeDepartments:
          type: number
        planType:
          type: string
        maxAgents:
          type: number
      required:
        - totalUsers
        - activeAgents
        - totalDepartments
        - activeDepartments
        - planType
        - maxAgents
...
    TenantUser:
      type: object
      properties:
        clerkUserId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, supervisor, agent]
        status:
          type: string
          enum: [active, invited, suspended]
        departments:
          type: array
          items:
            type: string
        joinedAt:
          type: string
          format: date-time
      required:
        - clerkUserId
        - email
        - role
        - status
        - departments
        - joinedAt
    UserInviteInput:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        departmentIds:
          type: array
          items:
            type: number
      required:
        - email
        - role
    TenantUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, supervisor, agent]
        status:
          type: string
          enum: [active, suspended]
...